add quantity in detail_product

+++ apps/store/models/models_product.py
@@ -7,7 +7,7 @@ class Product(models.Model):
     product_model = models.CharField(max_length=100)
     product_price = models.PositiveIntegerField()
     product_current_stock = models.PositiveIntegerField()
-    product_min_stock = models.DecimalField(max_digits=10, decimal_places=2)
+    product_min_stock = models.PositiveIntegerField()
     category = models.ForeignKey(Category, on_delete=models.CASCADE)
